Add addr argument to infcall_mmap
The addr argument of mmap was always set to 0 in infcall_mmap.
This allows setting it to another value, making it possible to
map specific pages in memory. This is useful when we need to jump
from the code to the page using a 32bit jump offset for instance,
as we can ask for a page within range.