Add 'patch' command to compile and inject code and x86_64 support
This command allows the user to dynamically compile code and inject it into the target process at a specific location so that it is executed within the program flow each time this specific instruction is hit.
In this version, only instruction with lengths superior to 5 bytes (jmp instruction) on x86_64 can be replaced.