Add the patch command to dynamically inject code in the inferior
The goal of this command is to compile and inject code in the inferior. 
This commit adds two commands : "patch code" and "patch file", which respectively take as input C/C++ code and a source file.
Compared to the compile command, these new commands leave the code in the binary, so that it can be executed several times.
The injection of code is done through a trampoline in a similar fashion to GDB fast tracepoints.
The original instruction is replaced with a jump to the trampoline which calls the gdb_expr() function generated by gcc (same process as in the compile command).
Changing computer